### 2021-11-05 #3894 @danielskatz
@noraloose - while we find an editor, can you work on the possibly missing DOI and the incorrect DOI that whedon suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references` to check again, and the command `@whedon generate pdf` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
