### 2021-11-05 #3894 @danielskatz
👋 @elbeejay - do you think you could edit this submission for JOSS?
