### 2021-11-05 #3894 @NoraLoose
Thanks @danielskatz for handling our submission and looking for an editor!

> while we find an editor, can you work on the possibly missing DOI and the incorrect DOI that whedon suggests

Yes, working on it!
