+++ paper.bib
@@ -30,7 +30,7 @@
 	title = {Dask: {Parallel} {Computation} with {Blocked} algorithms and {Task} {Scheduling}},
 	shorttitle = {Dask},
 	url = {http://conference.scipy.org/proceedings/scipy2015/matthew_rocklin.html},
-	doi = {10.25080/Majora-7b98e3ed-013},
+	doi = {10.25080/majora-7b98e3ed-013},
 	urldate = {2021-11-05},
 	journal = {Proceedings of the 14th Python in Science Conference},
 	author = {Rocklin, Matthew},
