### 2021-11-08 #3894 @elbeejay
@NoraLoose can you try switching the bib entry-type for `dask` to `@misc` instead of `@Manual`? That is the recommended format for citing a software repository per the JOSS docs ([link](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)), and so it might play nicer with whedon.
