### 2021-11-09 #3894 @arfon
Looks like @whedon was running out of memory, this can sometimes happen with very large vector images (such as [this one](https://github.com/ocean-eddy-cpt/gcm-filters/blob/master/filtered_vorticity.png)). If possible, please downsize that image, and this should be fine, but also, Whedon can likely be coaxed into compiling this (as just happened) 😸
