### 2021-11-09 #3894 @NoraLoose
Thanks @arfon for looking into this! I have decreased the figure size by half. Let's hope Whedon will like this paper better now!
