### 2021-11-10 #3894 @elbeejay
> Hi @elbeejay, I'd be happy to review this submission. I've not got much experience with GCMs, but can focus on the Python implementation, testing, documentation etc.
>
> I have not formal conflict of interest and I have never worked with the lead author. However, I am a contributor to the [glidertools](https://github.com/GliderToolsCommunity/GliderTools) package along with jbusecke, who has made several contributions to GCM-filters. I do not think this would affect my review of this submitssion, but please exclude me if this breaks the JOSS conflict of interest guidelines.

Hi @callumrollo. We'd appreciate your expertise on and review of the Pythonic elements of the submission. Thank you for being forthcoming about the potential `glidertools` COI, from your description it does not seem as if that runs afoul of the JOSS COI guidelines.
