### 2021-11-17 #3894 @elbeejay
Awesome @AleksiNummelin, sounds like you'll be able to provide perspectives as both a real user and fellow developer. Thank you for being clear about your relationship with the lead-author, I agree that it does not sound like a true COI.
