### 2021-11-22 #3894 @elbeejay
:wave: @isgiddy, would you be willing to review this submission to JOSS? I think your experience as an ocean scientist would allow you to provide the authors with a constructive review. At JOSS we do checklist-driven reviews, the JOSS review criteria can be found [here](https://joss.readthedocs.io/en/latest/review_criteria.html).

The submission is "GCM-Filters: A Python Package for Diffusion-based Spatial Filtering of Gridded Data". This is a pre-review issue which is used to find reviewers. Once 2 or 3 reviewers have been found I'll start the review on a dedicated GitHub issue. At present we are asking reviewers to complete reviews in 6 weeks, although this can be extended if needed.

If you are interested, please take a look at the journal's [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?#joss-conflict-of-interest-policy) to ensure you do not have a conflict before agreeing to review this submission.

Thank you, feel free to reach out if you have any questions about the JOSS review process. Please do not feel any pressure to accept this review request if you do not have the time or do not feel comfortable reviewing this software package, we appreciate and respect our peer-reviewers' time.
