### 2021-11-23 #3894 @elbeejay
> >
>
> Hi @elbeejay. This looks interesting and I'd be happy to learn more about this filtering technique and wider applications via the review process.

Awesome, thanks!
