### 2021-11-23 #3947 @whedon
[REVIEW]: GCM-Filters: A Python Package for Diffusion-based Spatial Filtering of Gridded Data
**Submitting author:** @NoraLoose (<a href="http://orcid.org/0000-0002-3684-9634">Nora Loose</a>)
**Repository:** <a href="https://github.com/ocean-eddy-cpt/gcm-filters" target ="_blank">https://github.com/ocean-eddy-cpt/gcm-filters</a>
**Version:** v0.2.3
