### 2021-11-23 #3947 @elbeejay
@callumrollo, @AleksiNummelin, and @isgiddy, thanks for agreeing to review this submission to JOSS. We are currently asking reviewers to try and complete their reviews in 6 weeks.

The JOSS review process is entirely open and transparent, and takes place on GitHub. Review comments can be made as issues in the **GCM-Filters** repository, please link this review issue when doing so (paste `https://github.com/openjournals/joss-reviews/issues/3947` into the issue).

For reference, here are links to the JOSS documentation that may be helpful as you conduct your reviews:

- [JOSS Review Criteria](https://joss.readthedocs.io/en/latest/review_criteria.html)
- [JOSS Review Checklist](https://joss.readthedocs.io/en/latest/review_checklist.html)

Please feel free to ping me (@elbeejay) if you have any questions/concerns. Thanks again for agreeing to review for JOSS.
