### 2021-11-30 de5c66e
Add Schmidt Futures acknowledgment to JOSS paper
