### 2021-12-02 #3947 @callumrollo
Hi Nora, this looks like a solid submission for JOSS. I'm not a modeller, but will review the code as best I can. I have a question about authorship. Most of the authors have made contributions to the codebase, corresponding to the author order. However, I could not find code contributions by  Laure
Zanna. Could you please outline their contribution to GCM_Filters?
