### 2021-12-03 #3947 @NoraLoose
Hi @callumrollo, thanks so much for reviewing our submission!

We defined the authorship criteria [here](https://github.com/ocean-eddy-cpt/gcm-filters/issues/44), following the JOSS recommendations.

The `GCM-Filters` package is being developed by the [Ocean Eddy Climate Process Team (CPT)](https://ocean-eddy-cpt.github.io/). (The CPT is also the github organization that owns the `GCM-Filters` repo.) Laure Zanna does not only lead the CPT, but has also made contributions through discussing code & package ideas and the implementation of the MOM5 Laplacians (partially off github's eye,  in our CPT meetings as well as in her research group meetings with Elizabeth Yankovsky and Arthur Guillaumin). She is invited as a co-author because her contributions clearly satisfy the following JOSS criterion:

> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are.

I hope this explanation helps! I'm happy to answer any further questions and concerns.
