### 2021-12-10 #3947 @callumrollo
There is a `notebooks` folder in the top level of the repo, with a jupyter notebook enticingly named `tutorial.ipynb`. However, these appears to be a demo for a lesson on packing or similar? Not sure what it's doing here. I'd suggest removing it lest potential users mistake it for a demo of the package functionality.

The documentation contains some great examples of package functionality. Perhaps these could be dumped into a notebook for ease of experimentation by potential users?
