### 2021-12-10 #3947 @callumrollo
Reading the API documentation, I am unclear on the difference between these two methods of the Filter class:

apply(self, field_or_dataset, dims):
"""Filter a field or xarray dataset with scalar Laplacian across the
dimensions specified by dims."""

apply_to_field(self, field, dims):
"""Filter a field with scalar Laplacian across the dimensions specified by dims."""

If `apply` can process a field or dataset, `apply_to_field` seems superfluous. If they are performing different functions, this could be better explained in their docstrings
