### 2021-12-10 #3947 @callumrollo
While not essential, the maintainers might consider copying the contributing information on the docs website to CONTRIBUTING.md in the repo so github detects it and users can find it more easily
