### 2021-12-10 #3947 @callumrollo
Similarly Issue and PR templates could make it easier for people to submit bug reports/requests etc. Again not an essential recommendation, it looks like the project is already dealing well with Issues and PRs
