### 2021-12-10 #3947 @callumrollo
Overall a very impressive package. Some of the workings are a bit beyond me, but I managed to run some of the example code and mess around with filter kernels. Great testing suite! Once above small issues are considered, I'm happy for this to go to publication
