### 2021-12-10 ocean-eddy-cpt/gcm-filters#110 @NoraLoose
API documentation for filter class methods
[As part of the JOSS review]( https://github.com/openjournals/joss-reviews/issues/3947#issuecomment-990591734), @callumrollo made a comment regarding our API documentation, which I am reposting here:

> Reading the API documentation, I am unclear on the difference between these two methods of the Filter class:
>
> apply(self, field_or_dataset, dims):
> """Filter a field or xarray dataset with scalar Laplacian across the
> dimensions specified by dims."""
>
> apply_to_field(self, field, dims):
> """Filter a field with scalar Laplacian across the dimensions specified by dims."""
>
> If apply can process a field or dataset, apply_to_field seems superfluous. If they are performing different functions, this could be better explained in their docstrings

@asross would you like to handle this issue? (I'm tagging you since you have worked on the filter class method related to xarray datasets. 😃) I think this only requires some minor fixes.
