### 2021-12-10 ocean-eddy-cpt/gcm-filters#113 @NoraLoose
Templates for Issues and PRs
@callumrollo [suggested](https://github.com/openjournals/joss-reviews/issues/3947#issuecomment-990598894) that we could have templates for Issues and PRs to make it easier for users to submit bug reports or requests. I like this idea, Callum! Will try to get to it soon.
