### 2021-12-10 ocean-eddy-cpt/gcm-filters#110 @asross
@NoraLoose sure, sounds good. Maybe the simplest fix is just to make `apply_to_field` a private method (which we just do by adding an `_` to its name? is that the convention?). Let me know if that makes sense.
