### 2021-12-10 ocean-eddy-cpt/gcm-filters#110 @asross
We could also merge them into a single `apply` function that recursively calls itself, I guess!
