### 2021-12-10 ocean-eddy-cpt/gcm-filters#110 @NoraLoose
> Maybe the simplest fix is just to make apply_to_field a private method (which we just do by adding an _ to its name? is that the convention?).

Making the method private (via a `_` prefix) sounds good to me. That's what we did in other places, for example [here](https://github.com/ocean-eddy-cpt/gcm-filters/blob/7dfd3fef4bacb912ae04cb1fadf3fcc839aba881/gcm_filters/kernels.py#L32-L33) or [here](https://github.com/ocean-eddy-cpt/gcm-filters/blob/7dfd3fef4bacb912ae04cb1fadf3fcc839aba881/gcm_filters/filter.py#L46).
