### 2021-12-10 2db051b
Make Filter#apply_to_field a private method and improve docstring of Filter#apply
