+++ gcm_filters/filter.py
@@ -412,14 +412,20 @@ class Filter:
         ax.legend()
 
     def apply(self, field_or_dataset, dims):
-        """Filter a field or xarray dataset with scalar Laplacian across the
-        dimensions specified by dims."""
+        """Filter a `field_or_dataset` (`xarray.DataArray` or `xarray.Dataset`)
+        with a scalar Laplacian across the dimensions specified by `dims`."""
+        if issubclass(self.Laplacian, BaseVectorLaplacian):
+            raise ValueError(
+                f"Provided Laplacian {self.Laplacian} is a vector Laplacian. "
+                f"The ``.apply`` method is only suitable for scalar Laplacians."
+            )
+
         if isinstance(field_or_dataset, xr.Dataset):
             filtered = field_or_dataset.copy(deep=True)
             any_filtered = False
             for key, field in filtered.variables.items():
                 if all(dim in field.dims for dim in dims):
-                    filtered[key] = self.apply_to_field(field, dims=dims)
+                    filtered[key] = self._apply_to_field(field, dims=dims)
                     any_filtered = True
             if not any_filtered:
                 warnings.warn(
@@ -429,16 +435,10 @@ class Filter:
                 )
             return filtered
         else:
-            return self.apply_to_field(field_or_dataset, dims=dims)
+            return self._apply_to_field(field_or_dataset, dims=dims)
 
-    def apply_to_field(self, field, dims):
+    def _apply_to_field(self, field, dims):
         """Filter a field with scalar Laplacian across the dimensions specified by dims."""
-        if issubclass(self.Laplacian, BaseVectorLaplacian):
-            raise ValueError(
-                f"Provided Laplacian {self.Laplacian} is a vector Laplacian. "
-                f"The ``.apply`` method is only suitable for scalar Laplacians."
-            )
-
         filter_func = _create_filter_func(self.filter_spec, self.Laplacian)
         grid_args = [self.grid_ds[name] for name in self.Laplacian.required_grid_args()]
         assert len(dims) == 2
