### 2021-12-11 #3947 @isgiddy
>

I've just started working my way through the documentation. It's very explicit and I've managed to work through all the examples and haven't run into any issues so far. The package is user friendly and a great contribution.

Some minor suggestions so far:

- I would suggest adding  `%time` to a few of the example filters that take a bit longer to run.

- there are a number of references in the docs, so perhaps adding a references.md page that collates them all would be a nice addition.

I'll aim to finish up going through everything this week
