### 2021-12-11 #3947 @elbeejay
Hi all, glad to see things are moving in a positive direction here.

A quick note to our reviewers @callumrollo, @isgiddy, and @AleksiNummelin; if you could make your comments and suggestions related to the review as issues in the [GCM-Filters repository](https://github.com/ocean-eddy-cpt/gcm-filters) itself that is our preference. This keeps those comments within the project repository should one go back through the commit history and PRs and look for *why* any particular changes were made. This can be done either by making individual issues for each comment, or by putting all of your comments related to the review into a single issue.

In the issue, please remember to just state that the issue is related to your JOSS review, and include the URL for this issue (https://github.com/ocean-eddy-cpt/gcm-filters), so that it is connected to this review issue.
