### 2021-12-14 7890cbb
Add options.extras_require to setup.cfg
