+++ docs/environment.yml
@@ -21,3 +21,9 @@ dependencies:
   - myst-parser
   - myst-nb
   - sphinxcontrib-srclinks
+  - cupy
+  - intake
+  - intake-xarray
+  - s3fs
+  - cartopy
+  - pooch >= 1.5.0
