### 2021-12-14 3ae79d0
Correct syntax for dependency version numbers
