+++ docs/environment.yml
@@ -26,4 +26,4 @@ dependencies:
   - intake-xarray
   - s3fs
   - cartopy
-  - pooch >= 1.5.0
+  - pooch>=1.5.0
