+++ setup.cfg
@@ -60,4 +60,4 @@ examples =
     intake-xarray
     s3fs
     cartopy
-    pooch >= 1.5.0
+    pooch>=1.5.0
