### 2021-12-14 1a81f28
Remove cupy from environment.yml
