+++ docs/environment.yml
@@ -21,7 +21,6 @@ dependencies:
   - myst-parser
   - myst-nb
   - sphinxcontrib-srclinks
-  - cupy
   - intake
   - intake-xarray
   - s3fs
