### 2021-12-14 62849e0
Re-add cupy to environment.yaml and comment out pooch
