### 2021-12-14 d824926
Re-add pooch and remove cupy from environment.yaml
