### 2021-12-15 #3947 @isgiddy
>

I've tested a few of the filter kernels now on an independent dataset. I'm happy for this to go to forward to publication.
