### 2021-12-17 #3947 @elbeejay
> Hi @elbeejay, sorry for being slow, I've started to look through the submission, but I had a simple question. I think I can't check the checkboxes above (should I be able to check them in the browser?), and I realized that I may not have accepted the invite and therefore I don't appear as an assignee, nor as a participant? Would it be possible to renew the invite, or is there something else that I am missing?

No problem, I will re-issue the review invitation.
