### 2021-12-18 fd92786
Fix tripole check for POP kernel
- test that northern edge grid data folds onto itself *only* away
  from land
