### 2021-12-20 ocean-eddy-cpt/gcm-filters#124 @AleksiNummelin
JOSS review - filtering on a tripolar grid and numerical instability
Dear @NoraLoose, all, this issue is related to my review for the JOSS submission https://github.com/openjournals/joss-reviews/issues/3947

Thanks for putting together this package, it is really nice addition to the general python universe related to ocean/climate modelling, and I'd expect some users to be interested in the methodology beyond these fields. I've tested the package on examples and some of my own data, and it mostly works as expected (see below). The documentation builds heavily on examples which is good I think, although that means that just based on API it might be a bit harder to understand how certain functions are working. Perhaps in time, one could link examples from the API (maybe this is unorthodox, but could help the user). Adding timing to some of the examples is a nice add on.

I only came across one issue that I could not make to work, and it might be related to https://github.com/ocean-eddy-cpt/gcm-filters/pull/123. When I tested the filtering on a [tripolar NorESM (BLOM) grid](http://ns9560k.web.sigma2.no/inputdata/ocn/blom/grid/grid_tnx0.25v4_20170622.nc), the code fails when checking that the southern most row is land. In the grid, that row is indeed land and `wet_mask[..., 0, :].any()` will give `False` although the check on line 438 gcm_filters/kernels.py somehow manages to give True. I didn't look much more into this, but I also wondered why this condition is needed, does the algorithm fail if it is violated? I think it is fine to publish the JOSS paper before this is fixed (might be my data), but would be good to solve this since the methodology should be general and one of the main benefits of the algorithm is really the possibility to filter on a tripolar grid.

Another small addition that I thought about was to show how the [numerical instability](https://gcm-filters.readthedocs.io/en/latest/examples/example_numerical_instability.html) is very sensitive to the number of grid cells. This seems to happen very abruptly at 30 grid cells for float32 and 66 grid cells for float128 using the MOM data. I wonder if these are general numbers independent of data? If so, maybe useful to mention and in any case might be better to show exactly when the instability occurs rather than showing the 10 degree case (which is far beyond what the failure point.)

I've attached [code](https://github.com/ocean-eddy-cpt/gcm-filters/files/7744862/gcm_filters_review_suggestions.txt) to this issue that goes trough these two cases (python code named as txt so that it can be attached here), hope it is useful, and thanks again for putting this together.
