### 2021-12-20 #3947 @AleksiNummelin
Okay, I've finally gone through the review and submitted an issue related to this at https://github.com/ocean-eddy-cpt/gcm-filters/issues/124

I think it is fine to publish the paper before that issue is fully resolved since those are rather minor points, although @NoraLoose might want to give a quick try solving the issue with filtering on a tripolar grid.
