### 2021-12-20 ocean-eddy-cpt/gcm-filters#124 @NoraLoose
Hi @AleksiNummelin,

Thanks for reviewing our JOSS submission and for your helpful comments!

> I only came across one issue that I could not make to work, and it might be related to #123. When I tested the filtering on a tripolar NorESM (BLOM) grid, the code fails when checking that the southern most row is land.

It's really great that you tested the tripolar Laplacian on the NorESM (BLOM) grid! You actually only have to make a minor modification to get your code to work:
```
delta_filtered_tripolar_regular_with_land = filter_tripolar_regular_with_land.apply(delta, dims=['y', 'x'])
```
rather than specifying `dims=['x',y']`. The order of the dimensions matters here. (I guess your code checked whether your *westernmost* row is equal to zeros.)
We should maybe emphasize in the documentation that the order of dimensions matters.

> I didn't look much more into this, but I also wondered why this condition is needed, does the algorithm fail if it is violated?

Yes, having land  in the southernmost row is a necessary condition. To see this, it is helpful to look at the following schematic of how the tripolar exchanges are coded:

![tripolar_exchanges](https://user-images.githubusercontent.com/23617395/146808223-5342cec9-874f-46ab-abf9-2e8c33af7876.png)

In step 1, if any of the southernmost row (U,V,W,X,Y,Z) is wet (i.e., not land), then the values in F,E,D,C,B,A will contaminate the values is U,V,W,X,Y,Z because we are applying diffusion with periodic boundary conditions. Does this make sense?
