### 2021-12-20 ocean-eddy-cpt/gcm-filters#124 @NoraLoose
@AleksiNummelin I will get back to your question about numerical instability in a bit. @iangrooms and @hmkhatri, feel free to jump in too.
