### 2021-12-20 f35d318
Apply suggestions from code review
Co-authored-by: Ryan Abernathey <ryan.abernathey@gmail.com>
