### 2021-12-20 29f9711
Update variable names and docstrings for consistency
