### 2021-12-20 ocean-eddy-cpt/gcm-filters#124 @AleksiNummelin
Thanks @NoraLoose, this clarifies! My mistake on the order of the dims then, I'll admit that I didn't read the API properly on this, I was just following the example, assuming that the order would not matter (didn't even think about it). I checked the API now, and it indeed does not specify that the order matters, maybe it would be helpful to add since the error doesn't really point to that - or at least I didn't come to think about that although it makes total sense now. The other option could be to flip the order if necessary, the only thing I could think of right now would be something like `if field.dims.index(dims[0])>field.dims.index(dims[1]): dims_new=dims[::-1]` which would work for the apply_to_field in filter.py and similarly one could use the ufield or vfield in apply_to_vector in filter.py, should work I think.
