### 2021-12-21 ocean-eddy-cpt/gcm-filters#124 @hmkhatri
Thanks @AleksiNummelin and @NoraLoose.

As mentioned in [Filter Theory](https://gcm-filters.readthedocs.io/en/latest/theory.html) section, roundoff errors lead to numerical instability. I am not sure if we could a priori identity the filter scale at which filtering scheme becomes unstable. Also, it is not clear if different types of fields, e.g. vectors vs tracers, would become unstable for the same filter scale.

The easiest way for a user is to try different filter scales and keep track of the total variance, which should decay with increasing the filter scale, in the filtered field. It should give an idea about the instability problem.
