### 2021-12-21 ocean-eddy-cpt/gcm-filters#124 @NoraLoose
@iangrooms came up with empirical values for when the filter becomes unstable:

https://github.com/ocean-eddy-cpt/gcm-filters/blob/75f172c7c74c1f2f3b91e36f7345bd138f01c881/gcm_filters/filter.py#L27-L36

For instance, in `n=2` dimensions and a Gaussian filter, the max filter factor is 77. If the user specifies a filter factor higher than that, the code will issue a warning:

https://github.com/ocean-eddy-cpt/gcm-filters/blob/75f172c7c74c1f2f3b91e36f7345bd138f01c881/gcm_filters/filter.py#L354-L357

But as @hmkhatri says and as we describe in [Grooms et al. (2021)](https://onlinelibrary.wiley.com/doi/abs/10.1029/2021MS002552), the maximum filter factor can't unfortunately be guaranteed a priori. It depends on the scales that are present in the field that the user wants to filter.
