### 2021-12-21 ocean-eddy-cpt/gcm-filters#124 @NoraLoose
> I checked the API now, and it indeed does not specify that the order matters, maybe it would be helpful to add since the error doesn't really point to that.

@AleksiNummelin, I totally agree that we have to do something about this issue. I opened #125.
