### 2021-12-21 #3947 @NoraLoose
Hi all!

Thanks for your comments and identifying issues - GCM-Filters is already seeing a lot of improvement thanks to your reviews! Most of our developers have already unplugged for the winter break, so I would like to resolve the outstanding issues and PRs in the first week of January.

I'm tracking the issues that you have raised and that still need to be resolved [here](https://github.com/ocean-eddy-cpt/gcm-filters/projects/1). (Everyone feel free to link more issues!)

Happy holidays, everyone!
