### 2021-12-22 #3947 @elbeejay
> Thanks for your comments and identifying issues - GCM-Filters is already seeing a lot of improvement thanks to your reviews! Most of our developers have already unplugged for the winter break, so I would like to resolve the outstanding issues and PRs in the first week of January.

@NoraLoose this sounds great, we will plan to revisit the submission in January in this case. Thanks @isgiddy,  @AleksiNummelin, and @callumrollo, it seems like this review process has been reasonably smooth.
