### 2022-01-05 9c0ca53
Switch RTD build to mamba
