+++ readthedocs.yml
@@ -1,9 +1,17 @@
-conda:
-    file: docs/environment.yml
+version: 2
+
+build:
+  os: "ubuntu-20.04"
+  tools:
+    python: "mambaforge-4.10"
+
 python:
-   version: 3.8
-   install:
-      - method: setuptools
-        path: package
+    install:
+       - method: setuptools
+         path: .
+
+conda:
+  environment: doc/environment.yml
+
 sphinx:
   fail_on_warning: true
