+++ readthedocs.yml
@@ -11,7 +11,7 @@ python:
          path: .
 
 conda:
-  environment: doc/environment.yml
+  environment: docs/environment.yml
 
 sphinx:
   fail_on_warning: true
