+++ readthedocs.yml
@@ -13,5 +13,5 @@ python:
 conda:
   environment: docs/environment.yml
 
-sphinx:
-  #fail_on_warning: true
+# sphinx:
+#   fail_on_warning: true
