+++ gcm_filters/filter.py
@@ -76,6 +76,7 @@ class FilterSpec(NamedTuple):
     is_laplacian: Iterable[bool]
     s_max: float
     p: Iterable[float]
+    n_iterations: int
 
 
 def _compute_filter_spec(
@@ -85,6 +86,7 @@ def _compute_filter_spec(
     transition_width=np.pi,
     ndim=2,
     n_steps=0,
+    n_iterations=1,
     root_tolerance=1e-8,
 ):
 
@@ -159,7 +161,7 @@ def _compute_filter_spec(
     s = np.array([y for x in s for y in x])
     is_laplacian = np.abs(s.imag / s.real) < root_tolerance
 
-    return FilterSpec(n_steps_total, s, is_laplacian, s_max, p)
+    return FilterSpec(n_steps_total, s, is_laplacian, s_max, p, n_iterations)
 
 
 def _create_filter_func(
@@ -184,19 +186,22 @@ def _create_filter_func(
         # filters, and does nothing for all other filters)
         field_bar = laplacian.prepare(field_bar)
 
-        for i in range(filter_spec.n_steps_total):
-            if filter_spec.is_laplacian[i]:
-                s_l = np.real(filter_spec.s[i])
-                tendency = laplacian(field_bar)  # Compute Laplacian
-                field_bar += (1 / s_l) * tendency  # Update filtered field
-            else:
-                s_b = filter_spec.s[i]
-                temp_l = laplacian(field_bar)  # Compute Laplacian
-                temp_b = laplacian(temp_l)  # Compute Biharmonic (apply Laplacian twice)
-                field_bar += (
-                    temp_l * 2 * np.real(s_b) / np.abs(s_b) ** 2
-                    + temp_b * 1 / np.abs(s_b) ** 2
-                )
+        for n in range(filter_spec.n_iterations):
+            for i in range(filter_spec.n_steps_total):
+                if filter_spec.is_laplacian[i]:
+                    s_l = np.real(filter_spec.s[i])
+                    tendency = laplacian(field_bar)  # Compute Laplacian
+                    field_bar += (1 / s_l) * tendency  # Update filtered field
+                else:
+                    s_b = filter_spec.s[i]
+                    temp_l = laplacian(field_bar)  # Compute Laplacian
+                    temp_b = laplacian(
+                        temp_l
+                    )  # Compute Biharmonic (apply Laplacian twice)
+                    field_bar += (
+                        temp_l * 2 * np.real(s_b) / np.abs(s_b) ** 2
+                        + temp_b * 1 / np.abs(s_b) ** 2
+                    )
 
         # finalize filtering (this divides by area for simple fixed factor filters,
         # and does nothing for all other filters)
@@ -230,30 +235,31 @@ def _create_filter_func_vec(
         # filters, and does nothing for all other filters)
         (ufield_bar, vfield_bar) = laplacian.prepare(ufield_bar, vfield_bar)
 
-        for i in range(filter_spec.n_steps_total):
-            if filter_spec.is_laplacian[i]:
-                s_l = np.real(filter_spec.s[i])
-                (utendency, vtendency) = laplacian(
-                    ufield_bar, vfield_bar
-                )  # Compute Laplacian
-                ufield_bar += (1 / s_l) * utendency  # Update filtered ufield
-                vfield_bar += (1 / s_l) * vtendency  # Update filtered vfield
-            else:
-                s_b = filter_spec.s[i]
-                (utemp_l, vtemp_l) = laplacian(
-                    ufield_bar, vfield_bar
-                )  # Compute Laplacian
-                (utemp_b, vtemp_b) = laplacian(
-                    utemp_l, vtemp_l
-                )  # Compute Biharmonic (apply Laplacian twice)
-                ufield_bar += (
[…]
