### 2022-01-10 10d6be5
Adds factorization of the Gaussian filter
