### 2022-01-11 01c128c
Test factored Gaussian
