### 2022-01-13 8705c41
Add eqn labels, fix more typos
