### 2022-01-13 c894c39
Re-organize new theory docs
