+++ docs/factored_gaussian.rst
@@ -1,0 +1,94 @@
+Factoring the Gaussian Filter
+=============================
+
+This section provides background on applying a Gaussian filter with a large filter scale by repeatedly applying Gaussian filters with a smaller filter scale.
+This was not discussed in `Grooms et al. (2021) <https://doi.org/10.1029/2021MS002552>`_, so this section provides extra detail.
+
+The :py:class:`gcm_filters.Filter` class has an argument ``n_iterations`` that will automatically split a Gaussian filter into a sequence of Gaussian filters with smaller scales, each of which is less sensitive to roundoff errors. If a user is encountering instability with the standard Gaussian filter, the user can try setting ``n_iterations`` to an integer greater than 1. In general ``n_iterations`` should be set to the smallest value that avoids numerical instability. The user should input their desired filter scale, and then the code will automatically adjust it to reflect the filter scale of the smaller constituent filters.
+Applying the Taper filter repeatedly is not equivalent to applying it once with a different scale, so this method only works for the Gaussian filter.
+
+.. note:: When ``n_iterations`` is greater than one, ``n_steps`` is the number of steps in a single small-scale Gaussian filter. The total number of steps taken by the algorithm is the product of ``n_iterations`` and ``n_steps``. If ``n_steps`` is not set by the user, or if it is set to a value less than 3, the code automatically changes ``n_steps`` to a default value that gives an accurate approximation to the small-scale Gaussian filters that comprise the factored filter.
+
+
+Mathematical Background
+-----------------------
+
+Per equation (13) of `Grooms et al. (2021) <https://doi.org/10.1029/2021MS002552>`_, the data to be filtered can be thought of as a vector :math:`\vec{f}` that can be expanded in an orthonormal basis of eigenfunctions of the discrete Laplacian
+
+.. math:: \vec{f} = \sum_i \hat{f}_i\vec{q}_i
+
+where :math:`\vec{q}_i` are the orthonormal basis vectors.
+Each basis vector is associated with a wavenumber :math:`k_i` and a length scale :math:`2\pi/k_i`.
+The filtered data can also be expanded in the same basis
+
+.. math:: \sum_i \hat{f}_ip(k_i^2)\vec{q}_i
+
+The polynomial :math:`p` approximates the target filter.
+If you filter the data :math:`N` times using the same filter, then the result has the following expansion
+
+.. math:: \sum_i \hat{f}_i(p(k_i^2))^N\vec{q}_i
+
+For a Gaussian filter, the target filter is
+
+.. math:: p(k^2) \approx g_L(k) = \text{exp}\left\{-\frac{L^2}{24}k^2\right\}
+
+where :math:`L` is the ``filter_scale``.
+The Gaussian target has the nice property that :math:`g_L^N = g_{\sqrt{N}L}`, i.e. if you apply the Gaussian filter :math:`N` times with scale :math:`L`, the result is the same as if you applied the Gaussian filter once with scale :math:`\sqrt{N}L`.
+The way the code uses this is that instead of filtering once with scale :math:`L`, it has the option to filter :math:`N` times with scale :math:`L/\sqrt{N}`.
+
+Inexact Equivalence
+-------------------
+
[…]
