+++ docs/index.rst
@@ -25,6 +25,7 @@ Contents
    examples/example_filter_types
    examples/example_tripole_grid
    examples/example_vector_laplacian
+   factored_gaussian
    examples/example_numerical_instability
    examples/example_satellite_observations
    api
