### 2022-01-14 9251a71
Iterated vector filter test
