### 2022-01-15 ocean-eddy-cpt/gcm-filters#124 @iangrooms
I'll second (third?) what's already been said about the numerical instability: It depends on a lot of factors, but most importantly it depends on the data being filtered which makes it practically impossible to predict a priori when instability will set in. The tests I ran (on which the warnings are based) use white noise, so the warnings should be "conservative" in the sense that the warning will hopefully kick in before the actual instability kicks in. Most people are filtering data that is nicer than white noise.

We're continually trying to improve this; there's a PR right now (#127) that significantly improves stability of the Gaussian filter.
