### 2022-01-17 7c77bd6
Create B grid function outline
