### 2022-01-20 c5cfc5f
Add downloads badge
