### 2022-01-21 ff9f12d
Correct Arthur's last name on JOSS paper
