### 2022-01-21 f339f06
Test n_iterations = 2,3,4
