### 2022-01-25 a067dac
Avoid NaN problems in northern edge folding check
