### 2022-02-02 5aed547
refactor vector test structure
