### 2022-02-02 6098884
update testing to include b-grid vector laplacian
