### 2022-02-02 3c423f9
further clean up tests
