### 2022-02-02 57959d5
add warning about dimension order to docs
