### 2022-02-03 a715297
handle NaNs in vector B-grid
