### 2022-02-03 ocean-eddy-cpt/gcm-filters#124 @NoraLoose
Hi @AleksiNummelin,

We have worked on addressing your two comments:

* **Order of filter dimensions**: #131 added warnings to the docs about that the dimension order matters. There are also some ideas in #125 for how to change the code in the long term to help address the problem of dimension order in a more automated fashion.
* **Numerical instability**: #127 and #130 added code and documentation for how to get around the numerical instability issue when using a Gaussian filter shape.

Let us know what you think, and thanks again for your comments!
