### 2022-02-03 ocean-eddy-cpt/gcm-filters#124 @AleksiNummelin
Thanks @NoraLoose, all, for the impressive work, this is really great! I think the documentation/warning approach on the dimension order is reasonable, I mean, we users should be able to write things in certain order now that it is clearly specified. The new information of the numerical instability in the notebook and especially the factor Gaussian approach are really neat. It is now clear to me and probably to a generic user as well what to expect in terms of the instability.

I'll add note to the review that my questions have been thoroughly answered - I think you can then close this issue.
