### 2022-02-03 #3947 @AleksiNummelin
Dear @elbeejay, @NoraLoose  and the team have made a really nice effort and answered my questions at https://github.com/ocean-eddy-cpt/gcm-filters/issues/124 so I consider that issue closed (left that for the authors), which means that this review is also complete from my side - happy to see this published.
