### 2022-02-03 #3947 @elbeejay
Cool, thanks @AleksiNummelin! @NoraLoose, let me know when you and your co-authors have made all of the revisions you'd like to make prior to JOSS publication. I noticed you created a "project" to track the issues which is neat. There's no timeline or deadline that you need to stick to at this point as all 3 reviewers have signed off, so I leave it up to you. Just ping me here when you'd like to move on to the final steps of the process.
