### 2022-02-03 ocean-eddy-cpt/gcm-filters#124 @NoraLoose
Thanks Aleksi!
