### 2022-02-03 890ded6
revert setup.cfg
