+++ setup.cfg
@@ -51,13 +51,3 @@ max-line-length = 105
 select = C,E,F,W,B,B950
 ignore = E203, E501, W503
 exclude = gcm_filters/_version.py
-
-[options.extras_require]
-examples =
-    matplotlib
-    cupy
-    intake
-    intake-xarray
-    s3fs
-    cartopy
-    pooch>=1.5.0
