### 2022-02-03 7691956
Change installation instructions to run notebook examples
