+++ docs/installation.rst
@@ -13,10 +13,6 @@ GCM-Filters can be installed via conda forge::
 
     conda install -c conda-forge gcm_filters
 
-If you want to run the example notebooks in this documentation, you will need a few extra dependencies that you can install via::
-
-    conda install -c conda-forge gcm_filters[examples]
-
 
 Installation from pip
 ^^^^^^^^^^^^^^^^^^^^^
@@ -27,10 +23,6 @@ GCM-Filters can also be installed with pip::
 
 This will install the latest release from
 `pypi <https://pypi.python.org/pypi>`_.
-
-The following command will install optional dependencies that are necessary to run the example notebooks in this documentation::
-
-    pip install gcm_filters[examples]
 
 
 Installation from GitHub
@@ -56,3 +48,13 @@ GCM-Filters and submit issues_ and `pull requests`_.
 .. _xarray: http://xarray.pydata.org
 .. _issues: https://github.com/ocean-eddy-cpt/gcm-filters/issues
 .. _`pull requests`: https://github.com/ocean-eddy-cpt/gcm-filters/pulls
+
+
+How to run the example notebooks
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+If you want to run the example notebooks in this documentation, you will need a few extra dependencies that you can install via::
+
+   mamba env create -f docs/environment.yml
+
+   conda activate gcm-filters-docs
