### 2022-02-05 64ec005
Introduce self.filter_scale_iterated for filter class
