### 2022-02-06 d4c38ec
Add code snippets to factored_gaussian.rst
