+++ docs/installation.rst
@@ -55,6 +55,6 @@ How to run the example notebooks
 
 If you want to run the example notebooks in this documentation, you will need a few extra dependencies that you can install via::
 
-   mamba env create -f docs/environment.yml
+   conda env create -f docs/environment.yml
 
    conda activate gcm-filters-docs
