### 2022-02-07 9c87f06
Fix plot_shape
