### 2022-02-08 #3947 @NoraLoose
Thanks @elbeejay! We have finished the last revisions (all tasks in the [JOSS project](https://github.com/ocean-eddy-cpt/gcm-filters/projects/1) are moved to "Done"). So I think we are ready to move on to the final steps.

Before doing so, I would like to make a small request. Between when we started the review and now, Paige Martin (@paigem) has started to work on PR https://github.com/ocean-eddy-cpt/gcm-filters/pull/128. The PR is not merged yet, but Paige has already made great progress toward an awesome new feature for our package! Therefore I have invited Paige as a co-author onto the JOSS paper. Sorry for the last-minute change, but I hope everyone here agrees with having Paige added to the author list.

Thanks again for your reviews and your super helpful feedback, @callumrollo, @isgiddy, and @AleksiNummelin!
