### 2022-02-10 #3947 @elbeejay
Hi @NoraLoose, I think the paper reads well. My only minor suggestions are to consider rewording/rephrasing in a few sections (see below) where `ndimage.gaussian_filter` and `GCM-Filters` end up being broken across lines. Up to you, but could be more aesthetically pleasing.

![image](https://user-images.githubusercontent.com/1770513/153453848-86383a16-edaa-4c2b-913e-2a519aebd87e.png)

![image](https://user-images.githubusercontent.com/1770513/153453892-d2af22b6-9b48-493e-ad36-ff03efd6b47b.png)

Whether or not you reword any parts of the paper, I'm happy for us to continue along the JOSS publication process. So at this time I'd like to ask if you could:

Doing so will enable us to square away the necessary metadata before this is published.
