+++ paper.md
@@ -78,7 +78,7 @@ Finally, the filter shape -- currently: either Gaussian or Taper -- determines h
 
 # Statement of Need
 
-Spatial filtering is commonly used as a scientific tool for analyzing gridded data. An example of an existing spatial filtering in python is the `ndimage.gaussian_filter` function in `SciPy` [@2020SciPy-NMeth], implemented as a sequence of convolution filters. While being a valuable tool for image processing (or blurring), `SciPy`'s Gaussian filter is of limited use for GCM data; it assumes a regular and rectangular Cartesian grid, employs a simple boundary condition, and the definitions of filter scale and shape have little or no flexibility. The python package `GCM-Filters` is specificially designed to filter GCM data, and seeks to solve a number of challenges for the user:
+Spatial filtering is commonly used as a scientific tool for analyzing gridded data. An example of an existing spatial filtering tool in python is the `ndimage.gaussian_filter` function in `SciPy` [@2020SciPy-NMeth], implemented as a sequence of convolution filters. While being a valuable tool for image processing (or blurring), `SciPy`'s Gaussian filter is of limited use for GCM data; it assumes a regular and rectangular Cartesian grid, employs a simple boundary condition, and the definitions of filter scale and shape have little or no flexibility. The python package `GCM-Filters` is specificially designed to filter GCM data, and seeks to solve a number of challenges for the user:
 
 1. GCM data comes on irregular curvilinear grids with spatially varying grid-cell geometry.
 2. Continental boundaries require careful and special treatment when filtering ocean GCM output.
@@ -96,7 +96,7 @@ The main `GCM-Filters` class that the user will interface with is the `gcm_filte
 * Grid types with **scalar Laplacians** that can be used for filtering scalar fields, for example temperature or vorticity (see \autoref{fig1}). The currently implemented grid types are compatible with different ocean GCM grids including MOM5 [@mom5], MOM6 [@adcroft2019MOM6] and the POP2 [@pop2] tripole grid.
 * Grid types with **vector Laplacians** that can be used for filtering vector fields, for example horizontal velocity $(u,v)$. The currently implemented grid type is compatible with ocean GCM grids that use an Arakawa C-grid convention; examples include MOM6 [@adcroft2019MOM6] and the MITgcm [@mitgcm].
 
-Atmospheric model grids are not yet supported, but could be implemented within `GCM-Filters`. Users are encouraged to contribute more grid types and Laplacians via pull requests.
+Atmospheric model grids are not yet supported, but could be implemented in `GCM-Filters`. Users are encouraged to contribute more grid types and Laplacians via pull requests.
 While we are excited to share GCM-Filters in its current beta state at version `0.1.3`, we plan to continue improving and maintaining the package for the long run and welcome new contributors from the broader community.
 
 # Acknowledgements
