### 2022-02-10 25e17eb
Update package version in JOSS paper
