### 2022-02-10 ae10797
Update release to 0.2.1 in JOSS paper
