### 2022-02-10 #3947 @NoraLoose
To be consistent between zenodo archive and JOSS paper, I changed v0.2 to v0.2.3 in the paper.
