### 2022-02-11 #3947 @elbeejay
Hi @NoraLoose, this looks good to me, I'll be recommending this for publication in JOSS; the final step is just going to involve one of the Editors-in-Chief looking over this review issue and the paper which should happen over the next few days. Thanks again to @AleksiNummelin, @callumrollo, and @isgiddy for volunteering your time and peer reviewing this submission.
