### 2022-02-11 #3947 @elbeejay
I'll note for the EiC that looks into this that there is no DOI associated with Dask; the reference used in this paper matches the citation recommended by Dask [here](https://docs.dask.org/en/latest/faq.html#how-do-i-cite-dask).
